refactor: decomposition keyboard.ts
