+++ src/languages/eng.ts
@@ -1,0 +1,6 @@
+export const boardEng = {
+  'Digit1': 'a',
+  'Digit2': 'b'
+}
+
+export default boardEng;
