+++ src/languages/rus.ts
@@ -1,0 +1,6 @@
+export const board = {
+  'Digit1': '1',
+  'Digit2': '2'
+}
+
+export default board;
